+++ CHANGES.rst
@@ -5,6 +5,7 @@
 6.2 (unreleased)
 ================
 
+- Add preliminary support for Python 3.13 as of 3.13a3.
 
 6.1 (2023-10-05)
 ================